move static package to test
