+++ main.go
@@ -5,7 +5,7 @@ import (
 
 	"tcp-load-balancer/internal/config"
 	"tcp-load-balancer/internal/server"
-	"tcp-load-balancer/internal/static"
+	"tcp-load-balancer/test"
 )
 
 func main() {
@@ -18,7 +18,7 @@ func main() {
 	log.Printf("Load balancer listening on %s", lb.Address())
 
 	// Manually configure upstream hosts and downstream clients to demonstrate functionality.
-	if err = static.Setup(lb, config.NumberOfHosts, config.NumberOfClients, config.ClientMessageInterval); err != nil {
+	if err = test.Setup(lb, config.NumberOfHosts, config.NumberOfClients, config.ClientMessageInterval); err != nil {
 		log.Fatalf("unable to setup static connection simulators: %s", err)
 	}
