+++ test/client.go
@@ -1,4 +1,4 @@
-package static
+package test
 
 import (
 	"log"
