+++ test/setup.go
@@ -1,9 +1,9 @@
-package static
+package test
 
-//////////////////////////////////////////////////////////////////////////////////////////////
-// This package is used entirely for demonstration purposes (configuring hosts and clients).
+////////////////////////////////////////////////////////////////////////////////////////////////////
+// This package is used entirely for demonstration purposes (to statically configure hosts/clients).
 // It can be removed in the future, once the LB has real hosts and clients to work with.
-//////////////////////////////////////////////////////////////////////////////////////////////
+////////////////////////////////////////////////////////////////////////////////////////////////////
 
 import (
 	"time"
